% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publicwww-package.R
\docType{package}
\name{publicwww}
\alias{publicwww}
\alias{publicwww-package}
\title{Query the 'PublicWWW' Source Code Search Engine}
\description{
The 'PublicWWW' \url{https://publicwww.com/} source code search engine indexes the content
of over 200 million web sites and provides a query interface that lets the
caller find any alphanumeric snippet, signature or keyword in the web pages
'HTML', 'JavaScript' and 'CSS' style sheet code.
}
\details{
The API key can be provided at the time of calling, but it is recommended that
the caller store it in the \code{PUBLICWWW_API_KEY} environment variable
(a good place for that is \code{~/.Renviron}).
}
\note{
The site requires an \href{https://publicwww.com/pricing.html}{API key} and has limited
functionality under the free plan.
}
\author{
Bob Rudis (bob@rud.is)
}
