% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{source_code_search}
\alias{source_code_search}
\title{Find any alphanumeric snippet, signature or keyword in the web pages HTML, JS and CSS code in the PublicWWW database}
\usage{
source_code_search(query, snippets = FALSE,
  pw_api_key = Sys.getenv("PUBLICWWW_API_KEY"))
}
\arguments{
\item{query}{query string}

\item{snippets}{if \code{TRUE} then include snippets in the download}

\item{pw_api_key}{PublicWWW API KEY}
}
\value{
data frame (tibble)
}
\description{
The API is \emph{highly} restrictive, even for paid plans, but even the free plan provides some
insight into the makeup of the top 3 million web sites.
}
\details{
The API key can be provided at the time of calling, but it is recommended that
the caller store it in the \code{PUBLICWWW_API_KEY} environment variable
(a good place for that is \code{~/.Renviron}).
}
\examples{
# Sites using the WordPress theme "Twenty Sixteen":
source_code_search('"/wp-content/themes/twentysixteen/"')

# Sites running the coinhive JS miner:
source_code_search('"coin-hive.com/lib/coinhive.min.js"')
}
\references{
\href{https://publicwww.com/}{PublicWWW home}; \href{https://publicwww.com/syntax.html}{Query syntax}; \href{https://publicwww.com/pricing.html}{API Pricing}
}
